+++ src/views/wny/dashboard/Developer.vue
@@ -57,6 +57,11 @@
                   <b-badge>{{ leadStatusesSoftDeleted }}</b-badge>
                 </p>
               </a>
+              <a href="#/admin/stages/soft-deleted">
+                <p>Stages
+                  <b-badge>{{ stagesSoftDeleted }}</b-badge>
+                </p>
+              </a>
             </div>
           </b-card>
         </b-col>
@@ -84,6 +89,7 @@
         leadSourcesSoftDeleted: '',
         tradesSoftDeleted: '',
         leadStatusesSoftDeleted: '',
+        stagesSoftDeleted: '',
       }
     },
 
@@ -106,7 +112,8 @@
                 this.lsCategoriesSoftDeleted = response.data.data[5].count,
                 this.leadSourcesSoftDeleted = response.data.data[6].count,
                 this.tradesSoftDeleted = response.data.data[7].count,
-                this.leadStatusesSoftDeleted = response.data.data[8].count
+                this.leadStatusesSoftDeleted = response.data.data[8].count,
+                this.stagesSoftDeleted = response.data.data[9].count
             );
           });
     }